consultplace/requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.3
django-debug-toolbar==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
ipython==8.23.0
django_extensions==3.2.3
python-telegram-bot==21.1.1
celery==5.4.0
aiogram==3.5.0
requests==2.31.0
python-dotenv==1.0.1
setuptools==70.0.0
redis==5.0.5
gunicorn==20.1.0
psycopg2-binary>=2.9.3,<3.0
flake8==7.1.1
